PR feedback: move ErrorBoundary into utils
